refactor: Break down generic.rs into smaller, focused modules
Refactored the large generic.rs file (1483 lines) into a well-organized
module structure with clear separation of concerns:

- filter.rs (410 lines): Filter predicate extraction and evaluation
- aggregation.rs (440 lines): Generic single-table aggregation patterns
- group_by.rs (630 lines): Generic GROUP BY aggregation patterns
- mod.rs (40 lines): Module organization and public API

Benefits:
- Improved code navigation and maintainability
- Clear logical boundaries between components
- Each module stays under 650 lines (well below the 1483 line original)
- All tests pass with no regressions
- Public API preserved for backwards compatibility
